requirements.txt
schedule
JSON_minify
slackclient 
slackeventsapi